LPD-72635 export-import-service: Correct package
