+++ modules/apps/export-import/export-import-service/src/main/java/com/liferay/exportimport/internal/batch/engine/action/DependencyManagementExportTaskPostAction.java
@@ -3,12 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.exportimport.internal.action;
+package com.liferay.exportimport.internal.batch.engine.action;
 
 import com.liferay.batch.engine.BatchEngineTaskItemDelegate;
-import com.liferay.batch.engine.action.ImportTaskPreAction;
-import com.liferay.batch.engine.context.ImportTaskContext;
-import com.liferay.batch.engine.model.BatchEngineImportTask;
+import com.liferay.batch.engine.action.ExportTaskPostAction;
+import com.liferay.batch.engine.model.BatchEngineExportTask;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessorRegistryUtil;
 import com.liferay.exportimport.internal.lar.ExportImportDescriptorThreadLocal;
@@ -16,6 +15,7 @@ import com.liferay.exportimport.internal.lar.PortletDataContextThreadLocal;
 import com.liferay.exportimport.kernel.lar.ExportImportThreadLocal;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.vulcan.batch.engine.ExportImportVulcanBatchEngineTaskItemDelegate;
+import com.liferay.petra.function.UnsafeSupplier;
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.log.Log;
@@ -33,18 +33,18 @@ import org.osgi.service.component.annotations.Component;
 /**
  * @author Carlos Correa
  */
-@Component(service = ImportTaskPreAction.class)
-public class DependencyManagementImportTaskPreAction
-	implements ImportTaskPreAction {
+@Component(service = ExportTaskPostAction.class)
+public class DependencyManagementExportTaskPostAction
+	implements ExportTaskPostAction {
 
 	@Override
 	public void run(
-			BatchEngineImportTask batchEngineImportTask,
+			BatchEngineExportTask batchEngineExportTask,
 			BatchEngineTaskItemDelegate<?> batchEngineTaskItemDelegate,
-			ImportTaskContext importTaskContext, Object item)
+			Object item)
 		throws Exception {
 
-		if (!ExportImportThreadLocal.isImportInProcess()) {
+		if (!ExportImportThreadLocal.isExportInProcess()) {
 			return;
 		}
 
@@ -86,18 +86,20 @@ public class DependencyManagementImportTaskPreAction
 			return;
 		}
 
+		Method getPropertyValueMethod = null;
+		Method setPropertyValueMethod = null;
+
 		for (Map.Entry<String, String[]> entry : references.entrySet()) {
 			String fieldName = entry.getKey();
 
+			Class<?> itemClass = item.getClass();
+
 			Method getMethod = _fetchDeclaredMethod(
-				item.getClass(),
-				"get" + StringUtil.upperCaseFirstLetter(fieldName));
+				itemClass, "get" + StringUtil.upperCaseFirstLetter(fieldName));
 
-			Method getPropertyValueMethod = null;
-
-			if (getMethod == null) {
+			if ((getMethod == null) && (getPropertyValueMethod == null)) {
 				getPropertyValueMethod = _fetchDeclaredMethod(
-					item.getClass(), "getPropertyValue", String.class);
+					itemClass, "getPropertyValue", String.class);
 			}
 
 			if ((getMethod == null) && (getPropertyValueMethod == null)) {
@@ -111,10 +113,15 @@ public class DependencyManagementImportTaskPreAction
 			}
 
 			Method setMethod = _fetchDeclaredMethod(
-				item.getClass(), "setPropertyValue", String.class,
-				Object.class);
+				itemClass, "set" + StringUtil.upperCaseFirstLetter(fieldName),
+				UnsafeSupplier.class);
 
-			if (setMethod == null) {
+			if ((setMethod == null) && (setPropertyValueMethod == null)) {
+				setPropertyValueMethod = _fetchDeclaredMethod(
+					itemClass, "setPropertyValue", String.class, Object.class);
+			}
+
+			if ((setMethod == null) && (setPropertyValueMethod == null)) {
 				if (_log.isDebugEnabled()) {
 					_log.debug(
 						"Unable to find the set method for the field \"" +
@@ -154,10 +161,21 @@ public class DependencyManagementImportTaskPreAction
 						getPropertyValueMethod.invoke(item, fieldName));
 				}
 
-				setMethod.invoke(
-					item, fieldName,
-					exportImportContentProcessor.replaceImportContentReferences(
-						value, portletDataContext));
+				if (setMethod != null) {
+					setMethod.invoke(
+						item,
+						(UnsafeSupplier)() ->
+							exportImportContentProcessor.
+								replaceExportContentReferences(
+									value, portletDataContext));
+				}
+				else {
+					setPropertyValueMethod.invoke(
+						item, fieldName,
+						exportImportContentProcessor.
+							replaceExportContentReferences(
+								value, portletDataContext));
+				}
 			}
 		}
 	}
@@ -181,6 +199,6 @@ public class DependencyManagementImportTaskPreAction
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		DependencyManagementImportTaskPreAction.class);
+		DependencyManagementExportTaskPostAction.class);
 
 }
