+++ modules/apps/export-import/export-import-service/src/main/java/com/liferay/exportimport/internal/batch/engine/action/DependencyManagementImportTaskPreAction.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.exportimport.internal.action;
+package com.liferay.exportimport.internal.batch.engine.action;
 
 import com.liferay.batch.engine.BatchEngineTaskItemDelegate;
-import com.liferay.batch.engine.action.ExportTaskPostAction;
-import com.liferay.batch.engine.model.BatchEngineExportTask;
+import com.liferay.batch.engine.action.ImportTaskPreAction;
+import com.liferay.batch.engine.context.ImportTaskContext;
+import com.liferay.batch.engine.model.BatchEngineImportTask;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessorRegistryUtil;
 import com.liferay.exportimport.internal.lar.ExportImportDescriptorThreadLocal;
@@ -15,7 +16,6 @@ import com.liferay.exportimport.internal.lar.PortletDataContextThreadLocal;
 import com.liferay.exportimport.kernel.lar.ExportImportThreadLocal;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.vulcan.batch.engine.ExportImportVulcanBatchEngineTaskItemDelegate;
-import com.liferay.petra.function.UnsafeSupplier;
 import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.log.Log;
@@ -33,18 +33,18 @@ import org.osgi.service.component.annotations.Component;
 /**
  * @author Carlos Correa
  */
-@Component(service = ExportTaskPostAction.class)
-public class DependencyManagementExportTaskPostAction
-	implements ExportTaskPostAction {
+@Component(service = ImportTaskPreAction.class)
+public class DependencyManagementImportTaskPreAction
+	implements ImportTaskPreAction {
 
 	@Override
 	public void run(
-			BatchEngineExportTask batchEngineExportTask,
+			BatchEngineImportTask batchEngineImportTask,
 			BatchEngineTaskItemDelegate<?> batchEngineTaskItemDelegate,
-			Object item)
+			ImportTaskContext importTaskContext, Object item)
 		throws Exception {
 
-		if (!ExportImportThreadLocal.isExportInProcess()) {
+		if (!ExportImportThreadLocal.isImportInProcess()) {
 			return;
 		}
 
@@ -86,20 +86,18 @@ public class DependencyManagementExportTaskPostAction
 			return;
 		}
 
-		Method getPropertyValueMethod = null;
-		Method setPropertyValueMethod = null;
-
 		for (Map.Entry<String, String[]> entry : references.entrySet()) {
 			String fieldName = entry.getKey();
 
-			Class<?> itemClass = item.getClass();
-
 			Method getMethod = _fetchDeclaredMethod(
-				itemClass, "get" + StringUtil.upperCaseFirstLetter(fieldName));
+				item.getClass(),
+				"get" + StringUtil.upperCaseFirstLetter(fieldName));
 
-			if ((getMethod == null) && (getPropertyValueMethod == null)) {
+			Method getPropertyValueMethod = null;
+
+			if (getMethod == null) {
 				getPropertyValueMethod = _fetchDeclaredMethod(
-					itemClass, "getPropertyValue", String.class);
+					item.getClass(), "getPropertyValue", String.class);
 			}
 
 			if ((getMethod == null) && (getPropertyValueMethod == null)) {
@@ -113,15 +111,10 @@ public class DependencyManagementExportTaskPostAction
 			}
 
 			Method setMethod = _fetchDeclaredMethod(
-				itemClass, "set" + StringUtil.upperCaseFirstLetter(fieldName),
-				UnsafeSupplier.class);
+				item.getClass(), "setPropertyValue", String.class,
+				Object.class);
 
-			if ((setMethod == null) && (setPropertyValueMethod == null)) {
-				setPropertyValueMethod = _fetchDeclaredMethod(
-					itemClass, "setPropertyValue", String.class, Object.class);
-			}
-
-			if ((setMethod == null) && (setPropertyValueMethod == null)) {
+			if (setMethod == null) {
 				if (_log.isDebugEnabled()) {
 					_log.debug(
 						"Unable to find the set method for the field \"" +
@@ -161,21 +154,10 @@ public class DependencyManagementExportTaskPostAction
 						getPropertyValueMethod.invoke(item, fieldName));
 				}
 
-				if (setMethod != null) {
-					setMethod.invoke(
-						item,
-						(UnsafeSupplier)() ->
-							exportImportContentProcessor.
-								replaceExportContentReferences(
-									value, portletDataContext));
-				}
-				else {
-					setPropertyValueMethod.invoke(
-						item, fieldName,
-						exportImportContentProcessor.
-							replaceExportContentReferences(
-								value, portletDataContext));
-				}
+				setMethod.invoke(
+					item, fieldName,
+					exportImportContentProcessor.replaceImportContentReferences(
+						value, portletDataContext));
 			}
 		}
 	}
@@ -199,6 +181,6 @@ public class DependencyManagementExportTaskPostAction
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		DependencyManagementExportTaskPostAction.class);
+		DependencyManagementImportTaskPreAction.class);
 
 }
